Move to vpu/common
