+++ inference-engine/src/vpu/graph_transformer/include/vpu/frontend/frontend.hpp
@@ -188,7 +188,7 @@ public:
 
     static CustomLayer::Ptr getSuitableCustomLayer(const std::vector<CustomLayer::Ptr>& customLayers, const ie::CNNLayerPtr&cnnLayer);
     static ie::ICNNNetwork::Ptr convertNetwork(ie::ICNNNetwork& network);
-    bool isLayerSupported(const std::string& type);
+    std::vector<std::string> getSupportedLayers();
 
 private:
     Data getVpuData(const ie::DataPtr& ieData) const;
